refactor(docs): flatten docs-kit and drop singleton wiring
