+++ apps/docs/src/docs-kit/resolve-from-catch-all.test.ts
