+++ apps/docs/src/docs-kit/resolve-from-catch-all.ts
