+++ apps/docs/src/layouts/docs-layout/index.ts
@@ -1,0 +1,1 @@
+export { DocsLayout as default, DocsLayout } from './docs-layout';
