--- apps/docs/src/lib/docs-kit/layout/docs-layout/index.ts
@@ -1,1 +1,0 @@
-export * from './docs-layout';
