--- apps/docs/src/lib/docs-kit/layout/index.ts
@@ -1,1 +1,0 @@
-export { DocsLayout as default, DocsLayout } from './docs-layout/docs-layout';
