+++ apps/docs/src/pages/docs/[...slug]/index.tsx
@@ -2,10 +2,9 @@ import { eco } from '@ecopages/core';
 import type { GetMetadata, GetStaticProps } from '@ecopages/core';
 import type { JsxRenderable } from '@ecopages/jsx';
 import { entries, getComponent, getEntryBySegments } from 'ecopages:content/docs';
-import '@/docs-kit.instance';
-import { getDocsMdxComponents } from '@/lib/docs-kit/mdx/docs-mdx-components';
-import { DocsLayout } from '@/lib/docs-kit/layout';
-import { resolveFromCatchAll } from '@/lib/docs-kit/navigation/resolve-from-catch-all';
+import { getDocsMdxComponents } from '@/docs-kit/mdx-components';
+import { resolveFromCatchAll } from '@/docs-kit/resolve-from-catch-all';
+import { DocsLayout } from '@/layouts/docs-layout';
 
 type DocsCatchAllProps = {
 	section: string;
